cmd: move snapshots utility main and options
