+++ cmd/dev/check_senders.cpp
@@ -26,12 +26,11 @@
 #include <silkworm/core/types/block_body_for_storage.hpp>
 #include <silkworm/core/types/evmc_bytes32.hpp>
 #include <silkworm/db/stages.hpp>
+#include <silkworm/infra/cmd/common.hpp>
 #include <silkworm/infra/common/decoding_exception.hpp>
 #include <silkworm/infra/common/directories.hpp>
 #include <silkworm/infra/common/log.hpp>
 #include <silkworm/infra/concurrency/signal_handler.hpp>
-
-#include "../common/common.hpp"
 
 using namespace silkworm;
 using namespace silkworm::db;
