+++ cmd/dev/db_toolbox.cpp
@@ -44,6 +44,7 @@
 #include <silkworm/db/genesis.hpp>
 #include <silkworm/db/prune_mode.hpp>
 #include <silkworm/db/stages.hpp>
+#include <silkworm/infra/cmd/common.hpp>
 #include <silkworm/infra/common/decoding_exception.hpp>
 #include <silkworm/infra/common/directories.hpp>
 #include <silkworm/infra/common/ensure.hpp>
@@ -53,8 +54,6 @@
 #include <silkworm/infra/concurrency/signal_handler.hpp>
 #include <silkworm/infra/concurrency/spawn.hpp>
 #include <silkworm/infra/test_util/task_runner.hpp>
-
-#include "../common/common.hpp"
 
 namespace fs = std::filesystem;
 using namespace silkworm;
