+++ cmd/test/CMakeLists.txt
@@ -59,9 +59,9 @@ if(NOT SILKWORM_CORE_ONLY)
   target_link_libraries(ethereum PRIVATE silkworm_infra evmc::loader CLI11::CLI11 magic_enum::magic_enum)
 
   # BE&KV Tests
-  add_executable(backend_kv_test "backend_kv_test.cpp" "../common/shutdown_signal.hpp" "../common/shutdown_signal.cpp")
-  target_link_libraries(backend_kv_test PRIVATE silkworm_node CLI11::CLI11 magic_enum::magic_enum)
+  add_executable(backend_kv_test backend_kv_test.cpp)
+  target_link_libraries(backend_kv_test PRIVATE silkworm_infra_cmd silkworm_node CLI11::CLI11 magic_enum::magic_enum)
 
-  add_executable(sentry_client_test "sentry_client_test.cpp")
+  add_executable(sentry_client_test sentry_client_test.cpp)
   target_link_libraries(sentry_client_test PRIVATE silkworm_sentry)
 endif()
