+++ cmd/test/backend_kv_test.cpp
@@ -34,13 +34,12 @@
 #include <silkworm/core/common/util.hpp>
 #include <silkworm/core/types/address.hpp>
 #include <silkworm/db/tables.hpp>
+#include <silkworm/infra/cmd/shutdown_signal.hpp>
 #include <silkworm/infra/common/log.hpp>
 #include <silkworm/infra/grpc/common/conversion.hpp>
 #include <silkworm/infra/grpc/common/util.hpp>
 #include <silkworm/interfaces/remote/ethbackend.grpc.pb.h>
 #include <silkworm/interfaces/remote/kv.grpc.pb.h>
-
-#include "../common/shutdown_signal.hpp"
 
 using namespace std::literals;
