+++ silkworm/infra/cmd/CMakeLists.txt
@@ -1,0 +1,26 @@
+#[[
+   Copyright 2022 The Silkworm Authors
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+]]
+
+set(TARGET silkworm_infra_cmd)
+
+find_package(Boost REQUIRED COMPONENTS headers)
+find_package(CLI11 REQUIRED)
+
+file(GLOB_RECURSE SRC CONFIGURE_DEPENDS "*.cpp" "*.hpp")
+
+add_library(${TARGET} ${SRC})
+
+target_link_libraries(${TARGET} PUBLIC silkworm_core silkworm_infra Boost::headers CLI11::CLI11 silkworm-buildinfo)
